+++ ChangeLog.md
@@ -1,4 +1,6 @@
 # ChangeLog
 
-- more cleaning on random files
-- added global style, color, and values
+- extracted csv importer to separate file.
+- add night indicator for day view.
+- fixed view transition animations
+- general file cleanup.